FASE3_PrintedTextRecognition/requirements.txt
# Python 3.10+
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0
scikit-image>=0.21.0
scipy>=1.11.0
pillow>=10.0.0
matplotlib>=3.7.0
tqdm>=4.66.0
